game
A Labyrinth is a tree-based maze in Royal House of England where the royals often enjoyed themselves with their servants who would get in and find their way out of a labyrinth maze. 

Even today unsuspected tourists who visit these get lost and one can often hear cries for help. 

Leading us to this project where we feel the computer Linux based version will be highly safer! 

An electronic version based on logical ICs and EPROMs was created by Elektor a Dutch based magazine in 1984. A reloaded version based on microcontroller was also done in 2020. We will implement it based on a Linux Console.